shopping-basket-demo: Do not support GStreamer with OpenCV on smarc-rzg2l
OpenCV camera support is buggy on the smarc-rzg2l when going
through GStreamer. Make sure we use the default version for
OpenCV (v4.1.0) for the Smarc RZ/G2L board, and prevent OpenCV
from trying to use GStreamer as a temporary workaround.
Make also sure that the RZ/G2L platfrom is the only platform
to use then new .bbappend.

Signed-off-by: Fabrizio Castro <fabrizio.castro.jz@renesas.com>